# Conflicts:
#	backend/archunit_store/dc710435-6d17-4b08-b30d-4f44792cfcce
#	backend/archunit_store/dffd983e-3f38-4d3f-a8e5-37227dfb0c3f
#	backend/archunit_store/stored.rules
#	backend/pom.xml
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/dto/OrderDetailsDto.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/dto/TicketDetailsDto.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/endpoint/LoginEndpoint.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/endpoint/ShowEndpoint.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/endpoint/TicketEndpoint.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/endpoint/exceptionhandler/GlobalExceptionHandler.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/endpoint/mapper/ShowResponseMapper.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/persistence/dao/HallSpotDao.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/persistence/dao/ShowDao.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/persistence/dao/TicketDao.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/persistence/entity/HallSector.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/persistence/mapper/ShowMapper.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/persistence/mapper/TicketMapper.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/persistence/mapper/UserMapper.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/persistence/repository/TicketRepository.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/security/HashUtil.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/security/LoginCacheConfig.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/service/ShowService.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/service/TicketService.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/service/UserService.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/service/impl/ShowServiceImpl.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/service/impl/TicketInvalidationSchedulingService.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/service/impl/TicketServiceImpl.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/service/impl/UserServiceImpl.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/service/job/InvalidateReservationJob.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/service/validator/OrderValidator.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/service/validator/TicketValidator.java
#	backend/src/main/resources/application.yml
#	backend/src/main/resources/quartz.properties
#	backend/src/test/java/at/ac/tuwien/sepr/groupphase/backend/architecture/ArchTests.java
#	backend/src/test/java/at/ac/tuwien/sepr/groupphase/backend/datagenerator/DataGeneratorUtils.java
#	backend/src/test/java/at/ac/tuwien/sepr/groupphase/backend/datagenerator/TicketViewDataGenerator.java
#	backend/src/test/java/at/ac/tuwien/sepr/groupphase/backend/integrationtest/AuthenticationEndpointTest.java
#	backend/src/test/java/at/ac/tuwien/sepr/groupphase/backend/integrationtest/OrderEndpointTest.java
#	backend/src/test/java/at/ac/tuwien/sepr/groupphase/backend/service/impl/OrderServiceImplTest.java
#	backend/src/test/java/at/ac/tuwien/sepr/groupphase/backend/service/impl/TicketInvalidationSchedulingServiceTest.java
#	backend/src/test/java/at/ac/tuwien/sepr/groupphase/backend/service/impl/TicketServiceImplTest.java
#	backend/src/test/java/at/ac/tuwien/sepr/groupphase/backend/supplier/HallPlanSupplier.java
#	backend/src/test/resources/quartz.properties
#	frontend/src/app/app-routing.module.ts
#	frontend/src/app/components/administrative-tasks/autocomplete-textfield/autocomplete-textfield.component.ts
#	frontend/src/app/components/administrative-tasks/create-show/create-show.component.ts
#	frontend/src/app/components/administrative-tasks/tag-search/tag-search.component.ts
#	frontend/src/app/components/user/manage-user/manage-user.component.ts
#	frontend/src/app/components/user/user-cart/user-cart.component.html
#	frontend/src/app/components/user/user-cart/user-cart.component.ts
#	frontend/src/app/services/openapi/.openapi-generator/FILES
#	frontend/src/app/services/openapi/model/models.ts
#	frontend/src/app/services/openapi/model/page-application-user-response.ts
#	frontend/src/app/services/openapi/model/page-location-dto.ts
#	frontend/src/app/services/openapi/model/page-show-response.ts
#	frontend/src/app/services/openapi/model/pageable-object.ts
#	frontend/src/app/services/openapi/model/show-dto.ts

+++ backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/service/impl/TicketServiceImpl.java
@@ -1,6 +1,9 @@
 package at.ac.tuwien.sepr.groupphase.backend.service.impl;
 
+import at.ac.tuwien.sepr.groupphase.backend.dto.ApplicationUserDto;
+import at.ac.tuwien.sepr.groupphase.backend.dto.OrderDetailsDto;
 import at.ac.tuwien.sepr.groupphase.backend.dto.OrderSummaryDto;
+import at.ac.tuwien.sepr.groupphase.backend.dto.TicketAddToOrderDto;
 import at.ac.tuwien.sepr.groupphase.backend.dto.TicketDetailsDto;
 import at.ac.tuwien.sepr.groupphase.backend.persistence.dao.HallSectorDao;
 import at.ac.tuwien.sepr.groupphase.backend.persistence.dao.HallSpotDao;
@@ -12,17 +15,19 @@ import at.ac.tuwien.sepr.groupphase.backend.security.HashUtil;
 import at.ac.tuwien.sepr.groupphase.backend.service.HallSectorShowService;
 import at.ac.tuwien.sepr.groupphase.backend.service.TicketService;
 import at.ac.tuwien.sepr.groupphase.backend.service.exception.DtoNotFoundException;
+import at.ac.tuwien.sepr.groupphase.backend.service.exception.ForbiddenException;
 import at.ac.tuwien.sepr.groupphase.backend.service.exception.ValidationException;
 import at.ac.tuwien.sepr.groupphase.backend.service.validator.TicketValidator;
-import jakarta.transaction.Transactional;
 import org.quartz.SchedulerException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.lang.invoke.MethodHandles;
 import java.time.Instant;
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class TicketServiceImpl implements TicketService {
@@ -71,13 +76,6 @@ public class TicketServiceImpl implements TicketService {
         return tickets;
     }
 
-    @Override
-    public List<TicketDetailsDto> findForShowById(long showId) {
-        var tickets = ticketDao.findForShowById(showId);
-        tickets.forEach(this::loadSectorShowForTicket);
-        return tickets;
-    }
-
     public void loadSectorShowForTicket(TicketDetailsDto ticket) {
         try {
             // handle cyclic dependencies between sectors and shows that apply to each ticket
@@ -97,7 +95,7 @@ public class TicketServiceImpl implements TicketService {
         try {
             var ticket = ticketDao.findById(id);
             ticketValidator.validateForCancelReservation(ticket);
-            ticketDao.cancelReservedTicket(id);
+            ticketDao.cancelReservedTicket(ticket.getId());
         } catch (EntityNotFoundException e) {
             throw new DtoNotFoundException(e);
         }
@@ -108,35 +106,42 @@ public class TicketServiceImpl implements TicketService {
         ticketDao.invalidateAllTicketsForOrder(orderId);
     }
 
-    @Transactional
     @Override
-    public TicketDetailsDto addTicketToOrder(Long seatId, Long showId, Long orderId, boolean reservationOnly)
-        throws ValidationException {
-        TicketDetailsDto ticketDetailsDto = createTicket(seatId, showId, orderId, reservationOnly);
+    public TicketDetailsDto addTicketToOrder(TicketAddToOrderDto ticket, ApplicationUserDto user)
+        throws ValidationException, ForbiddenException {
+        try {
+            var order = orderDao.findById(ticket.orderId());
+            if (!Objects.equals(user.getId(), order.getCustomer().getId())) {
+                throw new ForbiddenException();
+            }
+        } catch (EntityNotFoundException e) {
+            throw new ForbiddenException();
+        }
+
+        TicketDetailsDto ticketDetailsDto = createTicket(ticket);
         try {
             ticketInvalidationSchedulingService.scheduleReservationInvalidationsForNewlyAddedTicket(ticketDetailsDto);
         } catch (SchedulerException e) {
-            throw new IllegalStateException("Could not schedule reservation invalidation job", e);
+            throw new IllegalStateException("Could not schedule reservation invalidation job" + e.getMessage(), e);
         }
         return ticketDetailsDto;
     }
 
 
-    @Transactional
-    protected TicketDetailsDto createTicket(Long seatId, Long showId, Long orderId, boolean reservationOnly)
+    private TicketDetailsDto createTicket(TicketAddToOrderDto createTicket)
         throws ValidationException {
         TicketDetailsDto ticket = new TicketDetailsDto();
-        ticketValidator.validateForCreate(showId, seatId, orderId);
+        ticketValidator.validateForCreate(createTicket);
         try {
-            ticket.setHallSpot(hallSpotDao.findById(seatId));
-            ticket.setShow(showDao.findById(showId));
-            ticket.setReserved(reservationOnly);
+            ticket.setHallSpot(hallSpotDao.findById(createTicket.spotId()));
+            ticket.setShow(showDao.findById(createTicket.showId()));
+            ticket.setReserved(createTicket.reservationOnly());
             ticket.setValid(false);
-            String hashData = seatId + ":" + showId + ":" + orderId + ":" + Instant.now().toString();
+            String hashData = createTicket.spotId() + ":" + createTicket.showId() + ":" + createTicket.orderId() + ":" + Instant.now().toString();
             String ticketHash = HashUtil.generateHMAC(hashData);
             ticket.setHash(ticketHash);
-            var refOrder = new OrderSummaryDto();
-            refOrder.setId(orderId);
+            var refOrder = new OrderDetailsDto();
+            refOrder.setId(createTicket.orderId());
             ticket.setOrder(refOrder);
             // Note: @Peter R. ich habe das bissl geändert weil ich auch ein findSummaryById hatte.
             // Anundfürsich war die SummaryDto gedacht mit aggregtate Daten gefüllt zu sein, mein Request im Repository-Interface
@@ -152,11 +157,12 @@ public class TicketServiceImpl implements TicketService {
     }
 
     @Override
+    @Transactional
     public void confirmTicket(TicketDetailsDto ticketDetailsDto) throws SchedulerException, DtoNotFoundException {
         if (ticketDetailsDto.isReserved()) {
             ticketInvalidationSchedulingService.rescheduleReservationInvalidationJobForConfirmedOrder(ticketDetailsDto);
         } else {
-            ticketInvalidationSchedulingService.cancelReservationInvalidationJob(ticketDetailsDto.getId());
+            ticketInvalidationSchedulingService.cancelReservationInvalidationJob(ticketDetailsDto.getHash());
         }
         ticketDetailsDto.setValid(true);
         try {
@@ -172,13 +178,49 @@ public class TicketServiceImpl implements TicketService {
     }
 
     @Override
-    public void changeTicketReserved(long ticketId, boolean setReserved) throws ValidationException {
-        ticketDao.changeTicketReserved(ticketId, setReserved);
+    public TicketDetailsDto findByHash(String ticketHash) throws DtoNotFoundException {
+        try {
+            return ticketDao.findByHash(ticketHash);
+        } catch (EntityNotFoundException e) {
+            throw new DtoNotFoundException(e);
+        }
     }
 
     @Override
-    public void deleteTicket(long ticketId) throws ValidationException {
+    public void changeTicketReserved(long ticketId, boolean setReserved, ApplicationUserDto user)
+        throws ValidationException, DtoNotFoundException, ForbiddenException {
+        TicketDetailsDto ticket;
         try {
+            ticket = ticketDao.findById(ticketId);
+        } catch (EntityNotFoundException e) {
+            throw new DtoNotFoundException(e);
+        }
+
+        try {
+            var order = orderDao.findById(ticket.getOrder().getId());
+            if (!order.getCustomer().getId().equals(user.getId())) {
+                throw new ForbiddenException();
+            }
+
+            ticketValidator.validateForChangeTicketReserved(ticketId, order);
+
+            ticketDao.changeTicketReserved(ticketId, setReserved);
+        } catch (EntityNotFoundException e) {
+            throw new IllegalStateException("Ticket has invalid order id");
+        }
+    }
+
+    @Override
+    public void deleteTicket(long ticketId, ApplicationUserDto user) throws ValidationException, ForbiddenException {
+        try {
+            var ticket = ticketDao.findById(ticketId);
+            var order = orderDao.findById(ticket.getOrder().getId());
+
+            if (!order.getCustomer().getId().equals(user.getId())) {
+                throw new ForbiddenException();
+            }
+
+            ticketValidator.validateForDelete(ticketId, order);
             ticketDao.deleteById(ticketId);
         } catch (EntityNotFoundException ignored) {
             // ignored
